# sampleblog/go/model_article_response.go
/*
 * Swagger Blog
 *
 * A Simple Blog
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ArticleResponse struct {

	Id int32 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Username string `json:"username"`
}