# sampleblog/go/model_comment.go
/*
 * Swagger Blog
 *
 * A Simple Blog
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Comment struct {

	Date string `json:"date"`

	Content string `json:"content"`

	Author string `json:"author"`

	ArticleId int32 `json:"articleId"`
}