# sampleblog/go/model_comments.go
/*
 * Swagger Blog
 *
 * A Simple Blog
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Comments struct {

	Contents []Comment `json:"contents,omitempty"`
}