# sampleblog/go/model_user.go
/*
 * Swagger Blog
 *
 * A Simple Blog
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type User struct {

	Username string `json:"username"`

	Email string `json:"email,omitempty"`

	Password string `json:"password"`
	
	ID int `json:"id"`
}

type LUser struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

type SUser struct {
	Username string `json:"username"`
	Email string `json:"email"`
	ID int `json:"id"`
}

type AUsers struct {
	Contents []SUser `json:"contents,omitempty"`
}